add documentation of Python API
