+++ AMDirT/__init__.py
@@ -1,10 +1,1 @@
 __version__ = "1.4"
-import logging
-
-logger = logging.getLogger("AMDirT")
-logger.setLevel(logging.INFO)
-ch = logging.StreamHandler()
-ch.setLevel(logging.INFO)
-formatter = logging.Formatter("%(message)s")
-ch.setFormatter(formatter)
-logger.addHandler(ch)
