+++ docs/source/API.rst
@@ -1,0 +1,26 @@
+Python API
+==========
+
+Dataset Validation
+------------------
+
+.. autoclass:: AMDirT.validate.application.DatasetValidator
+    :members:
+    :show-inheritance:
+
+
+.. autoclass:: AMDirT.validate.application.AMDirValidator
+    :members:
+    :show-inheritance:
+
+Dataset conversion
+------------------
+
+.. autofunction:: AMDirT.convert.run_convert
+
+Dataset filtering
+------------------
+
+.. autofunction:: AMDirT.filter.run_app
+
+
