+++ tests/data/invalid_multi_archive_accession.tsv
@@ -1,0 +1,33 @@
+project_name	publication_year	publication_doi	site_name	latitude	longitude	geo_loc_name	sample_name	sample_host	sample_age	sample_age_doi	community_type	material	sampling_date	archive	archive_project	archive_accession
+Warinner2014	2014	10.1038/ng.2906	Dalheim	51.565	8.84	Germany	B61	Homo sapiens	900	10.1038/ng.2906	oral	dental calculus	NA	SRA	PRJNA216965	SRS473742,SRS473743,SRS473744,SRS473745
+Warinner2014	2014	10.1038/ng.2906	Dalheim	51.565	8.84	Germany	G12	Homo sapiens	900	10.1038/ng.2906	oral	dental calculus	NA	SRA	PRJNA216965	SRS473747,SRS473746,SRS473748,SRS473749,SRS473750
+Weyrich2017	2017	10.1038/nature21674	Gola Forest	7.657	-10.841	Sierra Leone	Chimp	Pan troglodytes	100	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890499
+Weyrich2017	2017	10.1038/nature21674	El Sidrón Cave	43.386	-5.328	Spain	ElSidron1	Homo sapiens neanderthalensis	49000	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890498
+Weyrich2017	2017	10.1038/nature21674	El Sidrón Cave	43.386	-5.329	Spain	ElSidron2	Homo sapiens neanderthalensis	49000	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890496
+Weyrich2017	2017	10.1038/nature21674	Spy Cave	50.48	4.674	Belgium	Spy1	Homo sapiens neanderthalensis	35800	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890491
+Weyrich2017	2017	10.1038/nature21674	Spy Cave	50.48	4.674	Belgium	Spy2	Homo sapiens neanderthalensis	36000	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890487
+Weyrich2017	2017	10.1038/nature21674	Dudka	53.96	21.86	Poland	EuroHG1	Homo sapiens	7600	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890494
+Weyrich2017	2017	10.1038/nature21674	Dudka	53.96	21.86	Poland	EuroHG2	Homo sapiens	7600	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890489
+Weyrich2017	2017	10.1038/nature21674	Stuttgart-Mühlhausen I	48.839	9.227	Germany	EuroLBK1	Homo sapiens	7400	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890488
+Weyrich2017	2017	10.1038/nature21674	Stuttgart-Mühlhausen I	48.839	9.227	Germany	EuroLBK2	Homo sapiens	7400	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890485
+Weyrich2017	2017	10.1038/nature21674	Stuttgart-Mühlhausen I	48.839	9.227	Germany	EuroLBK3	Homo sapiens	7400	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890490
+Weyrich2017	2017	10.1038/nature21674	Al Khiday	15.441	32.381	Sudan	Sudan1	Homo sapiens	5000	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890505
+Weyrich2017	2017	10.1038/nature21674	Al Khiday	15.441	32.381	Sudan	Sudan2	Homo sapiens	5000	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890503
+Weyrich2017	2017	10.1038/nature21674	Near Cape Town	NA	NA	South Africa	AfrSF1	Homo sapiens	2300	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890483
+Weyrich2017	2017	10.1038/nature21674	Near Cape Town	NA	NA	South Africa	AfrSF2	Homo sapiens	3800	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890484
+Weyrich2017	2017	10.1038/nature21674	Jewbury	53.962	-1.075	United Kingdom	UrbMed1	Homo sapiens	800	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890495
+Weyrich2017	2017	10.1038/nature21674	Jewbury	53.962	-1.075	United Kingdom	UrbWar2	Homo sapiens	800	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890492
+Weyrich2017	2017	10.1038/nature21674	Near Cape Town	NA	NA	South Africa	AfrPP2	Homo sapiens	600	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890486
+Weyrich2017	2017	10.1038/nature21674	Near Cape Town	NA	NA	South Africa	AfrPP1	Homo sapiens	600	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890504
+Weyrich2017	2017	10.1038/nature21674	Hettstedt Gymnasium	51.646	11.512	Germany	IndRev1	Homo sapiens	200	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890501
+Weyrich2017	2017	10.1038/nature21674	Hettstedt Gymnasium	51.646	11.512	Germany	IndRev2	Homo sapiens	200	10.1038/nature21674	oral	dental calculus	NA	SRA	PRJNA685265	SRS7890502
+Velsko2018	2018	10.1186/s40168-019-0717-3	Radcliffe Infirmary	51.752	-1.257	United Kingdom	CS01	Homo sapiens	200	10.1186/s40168-019-0717-3	oral	dental calculus	NA	ENA	PRJEB30331	ERS2985742,ERS2985790,ERS3881508
+Velsko2018	2018	10.1186/s40168-019-0717-3	Radcliffe Infirmary	51.752	-1.257	United Kingdom	CS02	Homo sapiens	200	10.1186/s40168-019-0717-3	oral	dental calculus	NA	ENA	PRJEB30331	ERS2985742,ERS2985790,ERS3881508
+Velsko2018	2018	10.1186/s40168-019-0717-3	Radcliffe Infirmary	51.752	-1.257	United Kingdom	CS03	Homo sapiens	200	10.1186/s40168-019-0717-3	oral	dental calculus	NA	ENA	PRJEB30331	ERS2985744,ERS2985792,ERS3881510
+Velsko2018	2018	10.1186/s40168-019-0717-3	Radcliffe Infirmary	51.752	-1.257	United Kingdom	CS04	Homo sapiens	200	10.1186/s40168-019-0717-3	oral	dental calculus	NA	ENA	PRJEB30331	ERS2985745,ERS2985745,ERS3881511
+Velsko2018	2018	10.1186/s40168-019-0717-3	Radcliffe Infirmary	51.752	-1.257	United Kingdom	CS05	Homo sapiens	200	10.1186/s40168-019-0717-3	oral	dental calculus	NA	ENA	PRJEB30331	ERS2985794,ERS3881512
+Velsko2018	2018	10.1186/s40168-019-0717-3	Radcliffe Infirmary	51.752	-1.257	United Kingdom	CS06	Homo sapiens	200	10.1186/s40168-019-0717-3	oral	dental calculus	NA	ENA	PRJEB30331	ERS2985795,ERS3881513
+Velsko2018	2018	10.1186/s40168-019-0717-3	Radcliffe Infirmary	51.752	-1.257	United Kingdom	CS07	Homo sapiens	200	10.1186/s40168-019-0717-3	oral	dental calculus	NA	ENA	PRJEB30331	ERS2985746,ERS2985796,ERS3881514
+Velsko2018	2018	10.1186/s40168-019-0717-3	Radcliffe Infirmary	51.752	-1.257	United Kingdom	CS08	Homo sapiens	200	10.1186/s40168-019-0717-3	oral	dental calculus	NA	ENA	PRJEB30331	ERS2985747,ERS2985747,ERS3881515
+Velsko2018	2018	10.1186/s40168-019-0717-3	Radcliffe Infirmary	51.752	-1.257	United Kingdom	CS09	Homo sapiens	200	10.1186/s40168-019-0717-3	oral	dental calculus	NA	ENA	PRJEB30331	ERS2985748,ERS2985798,ERS3881516
+Velsko2018	2018	10.1186/s40168-019-0717-3	Radcliffe Infirmary	51.752	-1.257	United Kingdom	CS10	Homo sapiens	200	10.1186/s40168-019-0717-3	oral	dental calculus	NA	ENA	PRJEB30331	ERS2985749,ERS2985799,ERS3881517
