+++ tests/test_application.py
@@ -13,3 +13,17 @@ def test_project_multiple_doi(test_data_dir):
     assert valid.check_duplicate_dois() is True
 
     assert invalid.check_duplicate_dois() is False
+
+
+def test_multi_values(test_data_dir):
+    valid_dataset = os.path.join(test_data_dir, "valid.tsv")
+    invalid_multiple_values = os.path.join(
+        test_data_dir, "invalid_multi_archive_accession.tsv"
+    )
+    schema = os.path.join(test_data_dir, "schema.json")
+
+    valid = AMDirValidator(schema=schema, dataset=valid_dataset)
+    invalid = AMDirValidator(schema=schema, dataset=invalid_multiple_values)
+
+    assert valid.check_multi_values() is True
+    assert invalid.check_multi_values() is False
